--- src/main/java/sportsmatchingservice/user/dto/UserPostDto.java
@@ -1,24 +1,0 @@
-package sportsmatchingservice.user.dto;
-
-
-import lombok.Getter;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-
-@Getter
-public class UserPostDto {
-
-    @NotBlank(message = "이메일을 작성해주세요.")
-    @Email(message = "이메일 형식에 맞게 작성해주세요.")
-    String email;
-
-    @NotBlank(message = "비밀번호를 입력해주세요.")
-    String password;
-
-    @NotBlank(message = "닉네임을 입력해주세요.")
-    String nickname;
-
-    @NotBlank(message = "전화번호를 입력해주세요.")
-    String phoneNumber;
-}
